requirements.txt
discord.py~=2.1.0
python-dotenv~=0.19.0
pillow~=8.4.0 
psycopg2-binary==2.9.3
openai~=0.27.4